isc_dhcp: Move dimension filling and bring another parser
